+++ OniAccess/ModConfig.cs
@@ -15,6 +15,7 @@ namespace OniAccess {
 		public bool TemperatureBandEarcons { get; set; } = false;
 		public bool FollowMovementEarcons { get; set; } = false;
 		public bool FootstepEarcons { get; set; } = true;
+		public bool ScannerDirectionEarcons { get; set; } = false;
 
 		public float UtilityPresenceVolume { get; set; } = 1.0f;
 		public float PipeShapeVolume { get; set; } = 0.15f;
@@ -23,5 +24,6 @@ namespace OniAccess {
 		public float FlowSonificationVolume { get; set; } = 0.05f;
 		public float FollowMovementVolume { get; set; } = 0.11f;
 		public float FootstepVolume { get; set; } = 1.5f;
+		public float ScannerDirectionVolume { get; set; } = 0.15f;
 	}
 }